Update sdif-rs to use real SDIF library API
This commit updates sdif-rs to use the actual IRCAM SDIF C library functions
instead of the stub bindings. All code now compiles and all tests pass with
the real bundled SDIF library.

## Changes to sdif-sys

### build.rs - Updated stub bindings
- Added SdifStringToSignature() / SdifSignatureToString()
- Updated NVT functions: SdifFNameValueList() (not SdifNameValueTableGetTable)
- Updated matrix type API: SdifCreateMatrixType() + SdifPutMatrixType()
- Updated frame type API: SdifCreateFrameType() + SdifPutFrameType()

### src/lib.rs - Added helper functions
- Added string_to_signature_c() - safe wrapper for C library signature conversion
- Documents the difference between bit manipulation and C library approaches

## Changes to sdif-rs

### src/builder.rs - Updated to real SDIF API
- NVT: SdifNameValueTableGetTable() → SdifFNameValueList()
- Matrix types: Now uses SdifCreateMatrixType() + SdifPutMatrixType() pattern
- Frame types: Now uses SdifCreateFrameType() + SdifPutFrameType() pattern
- Signatures: Uses string_to_signature() helper instead of SdifSignatureConst()
- Fixed pointer mutability for SdifFrameTypePutComponent()

### src/frame.rs - Fixed stream ID reading
- Changed SdifFGetSignature() to SdifFCurrID()
- The real SdifFGetSignature() takes 2 parameters, not 1
- SdifFCurrID() is the correct function for getting current frame's stream ID

### src/frame_builder.rs - Fixed type mismatches
- Cast u32 to usize for SdifFWritePadding() parameter

## Key API Differences: Stub vs Real

| Function | Stub (Guessed) | Real SDIF Library |
|----------|----------------|-------------------|
| Get NVT list | SdifNameValueTableGetTable | SdifFNameValueList |
| Create matrix type | SdifMatrixTypeLPutSdifMatrixType | SdifCreateMatrixType + SdifPutMatrixType |
| Create frame type | SdifFrameTypeLPutSdifFrameType | SdifCreateFrameType + SdifPutFrameType |
| Get stream ID | SdifFGetSignature(handle) | SdifFCurrID(handle) |
| Create signature | SdifSignatureConst(a,b,c,d) | SdifStringToSignature(str) |

## Real SDIF Library Pattern

The SDIF library uses a "create-configure-put" pattern:
1. Create: SdifCreateMatrixType(signature, NULL)
2. Configure: SdifMatrixTypeInsertTailColumnDef(mtype, "Column")
3. Put: SdifPutMatrixType(table, mtype)

Our stub bindings incorrectly assumed a single function would do all three steps.

## Test Results

✅ sdif-sys: 5/5 tests passing
✅ sdif-rs: 19/19 tests passing
✅ Full workspace builds successfully
✅ All code works with real SDIF C library (version 3.11.7)

## Documentation

Added SDIF_REAL_API_UPDATE.md with:
- Complete API mapping (stub vs real)
- Lessons learned about SDIF library patterns
- Build and test verification results

## Impact

This commit completes the transition from stub bindings to a fully functional
SDIF library integration. Both reading (Phase 2) and writing (Phase 3) now work
with the actual IRCAM SDIF C library compiled from bundled source on ARM64 macOS.

+++ sdif-rs/src/frame_builder.rs
@@ -285,7 +285,7 @@ impl<'a> FrameBuilder<'a> {
         }
 
         // Write padding
-        SdifFWritePadding(handle, calculate_padding(data_bytes));
+        SdifFWritePadding(handle, calculate_padding(data_bytes) as usize);
 
         Ok(())
     }
